requirements.txt
# Dependências Python para extrator de imagens PDF
pikepdf>=8.0.0
Pillow>=10.0.0
